+++ database/migrations/2022_11_13_185834_create_detail_penjualans_table.php
@@ -17,7 +17,7 @@ class CreateDetailPenjualansTable extends Migration
             $table->id();
             $table->foreignId('penjualan_id')->constrained('penjualans');
             $table->foreignId('produk_id')->constrained('produks');
-            $table->integer('qty');
+            $table->float('qty');
             $table->string('ket', 50);
             $table->integer('disc');
             $table->decimal('jumlah', 15, 1);
